tutorai-app/requirements.txt
streamlit
openai-whisper
transformers
TTS
torch
torchaudio
numpy